# src/main/java/org/lorislab/mcp/server/rs/user/UserRestService.java
/*
 * Copyright 2019 lorislab.org.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.lorislab.mcp.server.rs.user;

import javax.ws.rs.*;
import javax.ws.rs.core.MediaType;
import java.util.List;
import java.util.Map;

@Path("user")
@Consumes(MediaType.APPLICATION_JSON)
@Produces(MediaType.APPLICATION_JSON)
public interface UserRestService {

    @GET
    public List<User> getAll();

    @GET
    @Path("{id}")
    public User getById(@PathParam("id") Long id);

    @POST
    @Path("{id}")
    public User saveUser(@PathParam("id") Long id, User user);

    @GET
    @Path("map")
    public Map<Long, User> getMapUsers();

}